refactor(core): extract streaming methods into AgentStreamingMixin
DeileAgent grew to ~2934 lines / ~70 methods carrying lifecycle,
parsing, tool orchestration, streaming, autonomy, persona integration
and bot adapters in one class. Split the streaming-pipeline methods
(process_input_stream, process_input_stream_chunks,
_stream_chat_with_tools, plus their private helpers) into
AgentStreamingMixin in core/agent_streaming.py — DeileAgent now
inherits the mixin so call sites remain unchanged. Smaller agent.py
focused on orchestration; streaming path lives in a file that names
its responsibility.
